refactor: extract utility function generateOptions
